+++ UI/Android/app/src/main/jni/decodeYUV.h
@@ -25,7 +25,7 @@ void initEncoder(int width, int height);
 void Java_com_legaocar_rison_android_util_NativeUtil_initJpegEncoder
         (JNIEnv *env, jclass, int width, int height);
 
-jlong Java_com_legaocar_rison_android_util_NativeUtil_compressYuvToJpeg
+jint Java_com_legaocar_rison_android_util_NativeUtil_compressYuvToJpeg
         (JNIEnv *env, jclass, jbyteArray byteYuv, jbyteArray byteJpg,
          int format, int quality, int width, int height);
