+++ UI/Android/app/src/main/jni/yuv2jpg.h
@@ -319,7 +319,7 @@ void buildSTDHuffTab(BYTE *nrCodes, BYTE *stdTab, HUFFCODE *huffCode);
 int startEncode(JPEGINFO *pJpgInfo, BYTE *lpYBuf, BYTE *lpUBuf, BYTE *lpVBuf,
                 int width, int height, BYTE *pOut, int nDataLen);
 
-int YUV2Jpg(BYTE *in_Y, BYTE *in_U, BYTE *in_V, int width, int height, int nStride, int quality,
-            BYTE *pOut, unsigned long *pnOutSize);
+int YUV2Jpg(BYTE *in_Y, BYTE *in_U, BYTE *in_V, int width, int height, int nStride,
+            int quality, BYTE *pOut, int *pnOutSize);
 
 #endif
